@@ -12,7 +12,6 @@
 #include "driver/geometry/matrix/matrix.h"
 #include "driver/geometry/vector/vector.h"
 #include "driver/geometry/triangle/triangle.h"
-//#include "driver/light/light.h"
 
 typedef struct rasterised_points rasterised_points_t;
 struct rasterised_points
@@ -64,28 +63,19 @@ public:
     bool is_background(int index_i, int index_j);
 };
 
-Point<int> rasterize_triangle(std::vector<std::vector<rasterised_points_t>> &rasterized_points,
-                                    Triangle<double>&triangle_normals,
-                                    Vector3D<int> &light_position, Point<double> &point_1, Point<double> &point_2, Point<double> &point_3,
-                                    QGraphicsScene *scene, std::vector<std::vector<QColor>> &colors);
-
 void calculate_equation_plane(plane_koeffs_t &plane_koeffs,
                                                  Point<double> point_1, Point<double> point_2,
                                                  Point<double>point_3);
-/*void calculate_depth_pixels(std::vector<std::vector<double>> &zbuffer_matrix, std::vector<std::vector<QColor>> &color_matrix,
-                                           std::vector<std::vector<Vector2D>> &rasterized_points, plane_koeffs_t &plane_koeffs,
-                                            Vector3D<int> &light_position, Vector3D<double> &normal);*/
 
 void calculate_depth_pixels(std::vector<std::vector<double>> &zbuffer_matrix,
                                             std::vector<std::vector<QColor>> &color_matrix,
                                             std::vector<std::vector<rasterised_points_t>> &rasterized_points,
                                             plane_koeffs_t &plane_koeffs, Vector3D<int> &light_position,
-                                            Triangle<double>&triangle_normals, Triangle<double>points_3d,
-                                            Point<int> middle_point);
+                                            Triangle<double>&triangle_normals, Triangle<double>points_3d);
 
 Point<int> rasterize_triangle(std::vector<std::vector<rasterised_points_t>> &rasterized_points,
                                     Triangle<double>&triangle_normals,
                                     Vector3D<int> &light_position, Point<double> &point_1, Point<double> &point_2, Point<double> &point_3,
-                                    QGraphicsScene *scene, std::vector<std::vector<QColor>> &colors);
+                                    std::unique_ptr<QGraphicsScene>&scene, std::vector<std::vector<QColor>> &colors);
 
 #endif
