@@ -5,12 +5,26 @@
 #include <QMainWindow>
 #include <QGraphicsScene>
 #include <QDebug>
+#include <QImage>
 
 #include "driver/geometry/point/point.h"
 #include "driver/geometry/figure/figure.h"
 #include "driver/geometry/line/line.h"
 #include "driver/invisible/zbuffer.h"
 #include "driver/geometry/vector/vector.h"
+
+typedef struct meta_data meta_data_t;
+struct meta_data
+{
+    int octaves;
+    double amplitude;
+    double frequency;
+    double exponent;
+    int frequency_x;
+    int frequency_y;
+    double gain;
+    double lacunarity;
+};
 
 class Landscape
 {
@@ -21,19 +35,22 @@ private:
     std::vector<std::vector<Vector3D<int>>> _normals_down_triangles;
     std::vector<std::vector<Vector3D<double>>> _shading_normals;
 
+    rotate_t _rotate_landscape_angles;
+    meta_data_t _meta_config;
+
 public:
     Landscape();
     ~Landscape();
 
     void form_landscape();
     void transform_points_to_screen();
-    void draw_landscape(ZBuffer &zbuffer, QGraphicsScene *scene, QImage *image);
-    void remove_invisible_lines(ZBuffer &zbuffer, QGraphicsScene *scene, Vector3D<int> light_position);
+    void draw_landscape(ZBuffer &zbuffer, std::unique_ptr<QGraphicsScene>&scene, std::unique_ptr<QImage>&image);
+    void remove_invisible_lines(ZBuffer &zbuffer, std::unique_ptr<QGraphicsScene>&scene, Vector3D<int> light_position);
 
     void output_landscape();
     void output_screen_landscape();
 
-    void rotate_landscape(rotate_t &rotate_angles);
+    void rotate_landscape(rotate_t &diff_rotate_angles);
 
     Point<double> &get_point(int index_i, int index_j);
     Point<double> &get_screen_point(int index_i, int index_j);
@@ -46,6 +63,12 @@ public:
     void output_normals();
     void output_shading_normals();
     Vector3D<double> find_shading_normals(std::vector<Vector3D<int>> &normals, int i, int j);
+
+    rotate_t get_rotate_angles();
+    void set_rotate_angles(int angle_x, int angle_y, int angle_z);
+    void set_meta_config(int octaves, double amplitude, double frequency,
+                                      double exponent, int frequency_x, int frequency_y,
+                                      double gain, double lacunarity);
 };
 
 #endif
