@@ -8,9 +8,12 @@
 #include <QPainter>
 #include <QColor>
 #include <QDebug>
+#include <QScreen>
+#include <memory>
 
 #include "driver/geometry/polygon/polygon.h"
 #include "driver/landscape/landscape.h"
+#include "driver/light/light.h"
 
 QT_BEGIN_NAMESPACE
 namespace Ui { class MainWindow; }
@@ -21,15 +24,18 @@ class MainWindow : public QMainWindow
     Q_OBJECT
 
     Ui::MainWindow *ui;
-    QGraphicsScene *scene;
-    QGraphicsView *view;
-    QImage *image;
-    QPixmap *pixmap;
+    std::unique_ptr<QGraphicsScene> _scene;
+    std::unique_ptr<QGraphicsView> _view;
+    std::unique_ptr<QImage> _image;
+    QPixmap pixmap;
+    //QGraphicsPixmapItem* q_pmap;
     Landscape landscape; //  Сделал лендскейп членом класса. Иначе этот объект умирал при выходе из конструктора
-
+    Light light;
+    ZBuffer zbuffer;
 public:
     MainWindow(QWidget *parent = nullptr);
     ~MainWindow();
+    void init_landscape();
 
 public slots:
     void rotate_landscape_x();
